requirements.txt
# PROMETHEUS WISDOM — Dependencies
# AI Companion for Humanity

# LLM Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# Vector Store
chromadb>=0.5.0

# Graph Database
neo4j>=5.0.0

# Web Interface
streamlit>=1.40.0

# REST API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0

# Visualization
plotly>=5.24.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Code Quality
ruff>=0.8.0
